requirements.txt
argon2-cffi==19.2.0
bcrypt==3.1.7
bokeh==1.4.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
colorama==0.4.3
Django==2.2.8
djangorestframework==3.11.0
future==0.18.2
gunicorn==20.0.4
httpie==2.0.0
idna==2.8
Jinja2==2.10.3
MarkupSafe==1.1.1
numpy==1.18.1
packaging==20.1
pandas==0.24.2
Pillow==6.2.1
pycparser==2.19
Pygments==2.5.2
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
requests==2.22.0
scipy==1.4.1
six==1.13.0
sqlparse==0.3.0
tornado==6.0.3
urllib3==1.25.8
